Implement first plugin draft
We extract all symbols and determine the bounding box for all symbols
with the help of svgpath.

We support only rect and path for now because that seems to be what
inkscape outputs (it automatically creates ellipses and circles and
as paths internally).

Add readme documentation and a simple test.